requirements.txt
Authlib==1.6.5
blinker==1.9.0
boto3==1.42.7
botocore==1.42.7
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.1
Werkzeug==3.1.4
